backend/requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
djoser
channels
channels-redis
daphne
openai==0.28
